+++ resources/css/JADBudgetV2/dashboard.css
@@ -4,6 +4,13 @@
     #dashboard {
         @apply w-full h-full flex flex-row ;
 
+        .dashboardContainer {
+            .transactionsContainer {
+                @apply grid grid-cols-3 ;
+            }
+        }
+
+
         .pannelLeft {
             @apply w-fit md:w-1/5 h-full p-2 bg-white transition-all ;
