+++ resources/js/JADBudgetV2/Vue/Components/UI/BoxSignin.vue
@@ -1,8 +1,9 @@
 <template>
-    <div class="box boxLogin">
+    <div class="box boxSignin">
         <h1>{{ title }}</h1>
 
         <InputField :class="{'loading-state': loading}" name="login" type="text" placeholder="Identifiant" v-model="credentials.login"></InputField>
+        <InputField :class="{'loading-state': loading}" name="email" type="email" placeholder="E-mail" v-model="credentials.email"></InputField>
         <InputField :class="{'loading-state': loading}" name="password" type="password" placeholder="Mot de passe" :toggle-type="true" v-model="credentials.password"></InputField>
 
         <div class="link">
@@ -10,15 +11,15 @@
             <span>•</span>
             <a href="#">Mentions légales</a>
             <span>•</span>
-            <a href="#" @click.prevent="$emit('toggle-signin')">Inscrivez-vous</a>
+            <a href="#" @click.prevent="$emit('toggle-signin')">Connectez-vous</a>
         </div>
-        
-        <Button class="w-full" :class="{'loading-state': loading}" label="Se connecter" :disabled="loading" @click="login"></Button>
+
+        <Button class="w-full" :class="{'loading-state': loading}" label="S'inscrire" :disabled="loading" @click="signin"></Button>
 
     </div>
 </template>
 
-<style scoped>
+<style>
     h1 {
         margin-bottom: 2rem;
     }
@@ -49,16 +50,14 @@
 </style>
 
 <script>
-import InputField from './InputField.vue';
-import Button from './Button.vue';
-import Separator from './Separator.vue';
-import { fetch_result, makeToast } from '../../utils.ts';
+import InputField from './../Forms/InputField.vue';
+import Button from './../Buttons/Button.vue';
+import { fetch_result, makeToast } from '../../../../utils.ts';
 
 export default {
     components: {
         InputField,
-        Button,
-        Separator
+        Button
     },
     props: {
         title: {
@@ -72,34 +71,44 @@ export default {
             loading: false,
             credentials: {
                 login: '',
+                email: '',
                 password: ''
             }
         }
     },
     methods: {
-        async login() {
+        async signin() {
             this.loading = true;
-
+            
             const data = {
                 _token: document.querySelector('meta[name=_token]').getAttribute('content'),
-                login: this.credentials.login,
+                name: this.credentials.login,
+                email: this.credentials.email,
                 password: this.credentials.password
             };
 
-            const url = '/JADBudget/login';
+            const url = '/JADBudgetV2/signinV2';
 
             try {
-                const result = await fetch_result(url, data);
-                makeToast("success.png", "Connexion réussie !", 1500, () => { window.location.href = "/JADBudgetV2/dashboard" });
-
+                await fetch_result(url, data);
+                makeToast("success.png", "Inscription réussie !");
+                
             } catch (error) {
-                if (error.errors) {
-                    const firstError = Object.values(error.errors)[0][0];
-                    makeToast("error.png", firstError);
-                } 
-                else {
-                    makeToast("error.png", "Identifiants invalides, veuillez réessayer.");
+                try {
+                    const errorResponse = JSON.parse(error.message);
+                    
+                    if (errorResponse.errors) {
+                        const firstError = Object.values(errorResponse.errors)[0][0];
+                        makeToast("error.png", firstError);
+                    } else {
+                        makeToast("error.png", "Une erreur est survenue. Veuillez réessayer.");
+                    }
+
+                } catch (parseError) {
+                    console.log(parseError);
+                    makeToast("error.png", "Une erreur est survenue. Veuillez réessayer.");
                 }
+                
             } finally {
                 this.loading = false;
             }
