+++ resources/js/JADBudgetV2/dashboard.js
@@ -1,5 +1,5 @@
 import { createApp } from "vue";
-import Dashboard from "./Vue/Dashboard.vue";
+import Dashboard from "./Vue/Views/Dashboard.vue";
 
 document.addEventListener('DOMContentLoaded', () => {
     const dashboardApp = createApp(Dashboard);
